+++ hknweb/settings/common.py
@@ -35,10 +35,10 @@ INSTALLED_APPS = [
     "hknweb.course_surveys",
     "hknweb.exams",
     "hknweb.indrel",
-    "hknweb.resume",
     "hknweb.serv",
     "hknweb.tours",
     "hknweb.reviewsessions",
+    "hknweb.studentservices",
     "hknweb.thread",
     "dal",  # must be before django.contrib.admin
     "dal_select2",  # must be before django.contrib.admin
