+++ hknweb/studentservices/forms.py
@@ -1,5 +1,5 @@
 from django import forms
-from hknweb.resume.models import Resume
+from hknweb.studentservices.models import Resume
 
 
 class DocumentForm(forms.ModelForm):
