+++ hknweb/studentservices/views.py
@@ -1,18 +1,23 @@
 from django.shortcuts import render
-from hknweb.resume.forms import DocumentForm
+
+from hknweb.studentservices.forms import DocumentForm
 
 
-def index(request):
+SUBMIT_TEMPLATE = "studentservices/resume_critique_submit.html"
+UPLOADED_TEMPLATE = "studentservices/resume_critique_uploaded.html"
+
+
+def resume_critique_submit(request):
     if request.method == "POST":
         form = DocumentForm(request.POST, request.FILES)
         if form.is_valid():
             form.save()
-            return render(request, "resume/uploaded.html")
+            return render(request, UPLOADED_TEMPLATE)
         else:
             form = DocumentForm()
             return render(
                 request,
-                "resume/index.html",
+                SUBMIT_TEMPLATE,
                 {
                     "form": form,
                     "err": True,
@@ -22,7 +27,7 @@ def index(request):
     form = DocumentForm()
     return render(
         request,
-        "resume/index.html",
+        SUBMIT_TEMPLATE,
         {
             "form": form,
             "err": False,
@@ -30,6 +35,6 @@ def index(request):
     )
 
 
-def submitted(request):
+def resume_critique_uploaded(request):
     form = DocumentForm()
-    return render(request, "resume/index.html", {"form": form})
+    return render(request, SUBMIT_TEMPLATE, {"form": form})
