+++ hknweb/templates/base.html
@@ -107,7 +107,7 @@
       <span class="slash">/</span>
       <a href="{% url 'tours:index' %}">Department Tours</a>
       <span class="slash">/</span>
-      <a href="{% url 'resume:index' %}">Resume Critique</a>
+      <a href="{% url 'studentservices:resume' %}">Resume Critique</a>
     </div>
     <div class="navi-container mobile-nav-bar" style="text-align: center;">
       <label for="toggle-industry">
@@ -197,8 +197,6 @@
           <a href="{% url 'candidate:officer' %}">Officer Portal</a>
           <span class="slash">/</span>
         {% endif %}
-        <a href="{% url 'resume:index' %}">Resume Review</a>
-        <span class="slash">/</span>
         <a href="/cand">Candidate Portal</a>
         <span class="slash">/</span>
         <a href="/accounts/settings">Settings</a>
@@ -222,7 +220,7 @@
         <span class="slash">/</span>
         <a href="{% url 'tours:index' %}">Department Tours</a>
         <span class="slash">/</span>
-        <a href="{% url 'resume:index' %}">Resume Critique</a>
+        <a href="{% url 'studentservices:resume' %}">Resume Critique</a>
       </div>
 
       <div class="heading">
